Errors of the last ones are corrected. The project is formed.

+++ src/main/java/Seminar_1/Units/BasicHero.java
@@ -14,7 +14,7 @@ public abstract class BasicHero implements Interface_BH {
     protected Coordinates position;
     protected int initiative; // инициатива по ходу
     protected States state; // статус персонажа
-    protected float curHp; // текущее здоровье
+    protected int curHp; // текущее здоровье
 
 
     public BasicHero(Names name, int hit, int health, int curHp, int stamina, int row, int col, int initiative) {
